+++ src/sisl/_core/geometry.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 import logging
 import warnings
 from collections import OrderedDict
-from functools import reduce, singledispatchmethod
+from functools import singledispatchmethod
 from itertools import product
 from math import acos
 from numbers import Integral, Real
@@ -75,7 +75,6 @@ from sisl.utils.mathematics import fnorm
 from .atom import Atom, Atoms
 from .lattice import Lattice, LatticeChild
 from .orbital import Orbital
-from .quaternion import Quaternion
 
 __all__ = ["Geometry", "sgeom", "AtomCategory"]
 
@@ -811,7 +810,7 @@ class Geometry(
 
     def __str__(self) -> str:
         """str of the object"""
-        s = self.__class__.__name__ + f"{{na: {self.na}, no: {self.no},\n "
+        s = f"{self.__class__.__name__}{{na: {self.na}, no: {self.no},\n "
         s += str(self.atoms).replace("\n", "\n ")
         if len(self.names) > 0:
             s += ",\n " + str(self.names).replace("\n", "\n ")
@@ -1248,14 +1247,6 @@ class Geometry(
             raise ValueError(
                 f"{self.__class__.__name__}.iter_block got unexpected 'method' argument: {method}"
             )
-
-    def copy(self) -> Geometry:
-        """Create a new object with the same content (a copy)."""
-        g = self.__class__(
-            np.copy(self.xyz), atoms=self.atoms.copy(), lattice=self.lattice.copy()
-        )
-        g._names = self.names.copy()
-        return g
 
     def overlap(
         self,
@@ -1913,28 +1904,6 @@ class Geometry(
 
         return nsc
 
-    def sub(self, atoms: AtomsArgument) -> Geometry:
-        """Create a new `Geometry` with a subset of this `Geometry`
-
-        Indices passed *MUST* be unique.
-
-        Negative indices are wrapped and thus works.
-
-        Parameters
-        ----------
-        atoms : int or array_like
-            indices/boolean of all atoms to be removed
-
-        See Also
-        --------
-        Lattice.fit : update the supercell according to a reference supercell
-        remove : the negative of this routine, i.e. remove a subset of atoms
-        """
-        atoms = self.sc2uc(atoms)
-        return self.__class__(
-            self.xyz[atoms, :], atoms=self.atoms.sub(atoms), lattice=self.lattice.copy()
-        )
-
     def sub_orbital(self, atoms: AtomsArgument, orbitals: OrbitalsArgument) -> Geometry:
         r"""Retain only a subset of the orbitals on `atoms` according to `orbitals`
 
@@ -2056,28 +2025,6 @@ class Geometry(
 
         return geom
 
-    def remove(self, atoms: AtomsArgument) -> Geometry:
-        """Remove atoms from the geometry.
-
-        Indices passed *MUST* be unique.
-
-        Negative indices are wrapped and thus works.
-
-        Parameters
-        ----------
-        atoms : int or array_like
-            indices/boolean of all atoms to be removed
-
-        See Also
-        --------
-        sub : the negative of this routine, i.e. retain a subset of atoms
-        """
-        atoms = self.sc2uc(atoms)
-        if atoms.size == 0:
-            return self.copy()
-        atoms = np.delete(_a.arangei(self.na), atoms)
-        return self.sub(atoms)
-
     def remove_orbital(
         self, atoms: AtomsArgument, orbitals: OrbitalsArgument
     ) -> Geometry:
@@ -2127,160 +2074,6 @@ class Geometry(
 
         # now call sub_orbital
         return self.sub_orbital(atoms, orbitals)
-
-    def unrepeat(self, reps: int, axis: int, *args, **kwargs) -> Geometry:
-        """Unrepeats the geometry similarly as `untile`
-
-        Please see `untile` for argument details, the algorithm and arguments are the same however,
-        this is the opposite of `repeat`.
-        """
-        atoms = np.arange(self.na).reshape(-1, reps).T.ravel()
-        return self.sub(atoms).untile(reps, axis, *args, **kwargs)
-
-    def untile(
-        self,
-        reps: int,
-        axis: int,
-        segment: int = 0,
-        rtol: float = 1e-4,
-        atol: float = 1e-4,
-    ) -> Geometry:
-        """A subset of atoms from the geometry by cutting the geometry into `reps` parts along the direction `axis`.
-
-        This will effectively change the unit-cell in the `axis` as-well
-        as removing ``self.na/reps`` atoms.
-        It requires that ``self.na % reps == 0``.
-
-        REMARK: You need to ensure that all atoms within the first
-        cut out region are within the primary unit-cell.
-
-        Doing ``geom.untile(2, 1).tile(2, 1)``, could for symmetric setups,
-        be equivalent to a no-op operation. A ``UserWarning`` will be issued
-        if this is not the case.
-
-        This method may be regarded as the opposite of `tile`.
-
-        Parameters
-        ----------
-        reps :
-            number of times the structure will be cut (untiled)
-        axis :
-            the axis that will be cut
-        segment :
-            returns the i'th segment of the untiled structure
-            Currently the atomic coordinates are not translated,
-            this may change in the future.
-        rtol :
-            directly passed to `numpy.allclose`
-        atol :
-            directly passed to `numpy.allclose`
-
-        Examples
-        --------
-        >>> g = sisl.geom.graphene()
-        >>> gxyz = g.tile(4, 0).tile(3, 1).tile(2, 2)
-        >>> G = gxyz.untile(2, 2).untile(3, 1).untile(4, 0)
-        >>> np.allclose(g.xyz, G.xyz)
-        True
-
-        See Also
-        --------
-        tile : opposite method of this
-        """
-        if self.na % reps != 0:
-            raise ValueError(
-                f"{self.__class__.__name__}.untile "
-                f"cannot be cut into {reps} different "
-                "pieces. Please check your geometry and input."
-            )
-        # Truncate to the correct segments
-        lseg = segment % reps
-        # Cut down cell
-        lattice = self.lattice.untile(reps, axis)
-        # List of atoms
-        n = self.na // reps
-        off = n * lseg
-        new = self.sub(_a.arangei(off, off + n))
-        new.set_lattice(lattice)
-        if not np.allclose(new.tile(reps, axis).xyz, self.xyz, rtol=rtol, atol=atol):
-            warn(
-                "The cut structure cannot be re-created by tiling\n"
-                "The tolerance between the coordinates can be altered using rtol, atol"
-            )
-        return new
-
-    def repeat(self, reps: int, axis: int) -> Geometry:
-        """Create a repeated geometry
-
-        The atomic indices are *NOT* retained from the base structure.
-
-        The expansion of the atoms are basically performed using this
-        algorithm:
-
-        >>> ja = 0
-        >>> for ia in range(self.na):
-        ...     for id,r in args:
-        ...        for i in range(r):
-        ...           ja = ia + cell[id,:] * i
-
-        For geometries with a single atom this routine returns the same as
-        `tile`.
-
-        Tiling and repeating a geometry will result in the same geometry.
-        The *only* difference between the two is the final ordering of the atoms.
-
-        Parameters
-        ----------
-        reps :
-           number of repetitions
-        axis :
-           direction of repetition, 0, 1, 2 according to the cell-direction
-
-        Examples
-        --------
-        >>> geom = Geometry([[0, 0, 0], [0.5, 0, 0]], lattice=1)
-        >>> g = geom.repeat(2,axis=0)
-        >>> print(g.xyz) # doctest: +NORMALIZE_WHITESPACE
-        [[0.   0.   0. ]
-         [1.   0.   0. ]
-         [0.5  0.   0. ]
-         [1.5  0.   0. ]]
-        >>> g = geom.repeat(2,0).repeat(2,1)
-        >>> print(g.xyz) # doctest: +NORMALIZE_WHITESPACE
-        [[0.   0.   0. ]
-         [0.   1.   0. ]
-         [1.   0.   0. ]
-         [1.   1.   0. ]
-         [0.5  0.   0. ]
-         [0.5  1.   0. ]
-         [1.5  0.   0. ]
-         [1.5  1.   0. ]]
-
-        See Also
-        --------
-        tile : equivalent but different ordering of final structure
-        """
-        if reps < 1:
-            raise ValueError(
-                f"{self.__class__.__name__}.repeat requires a repetition above 0"
-            )
-
-        lattice = self.lattice.repeat(reps, axis)
-
-        # Our first repetition *must* be with
-        # the former coordinate
-        xyz = np.repeat(self.xyz, reps, axis=0)
-        # We may use broadcasting rules instead of repeating stuff
-        xyz.shape = (self.na, reps, 3)
-        nr = _a.arangei(reps)
-        nr.shape = (1, reps)
-        for i in range(3):
-            # Correct the unit-cell offsets along `i`
-            xyz[:, :, i] += nr * self.cell[axis, i]
-        xyz.shape = (-1, 3)
-
-        # Create the geometry and return it
-        return self.__class__(xyz, atoms=self.atoms.repeat(reps), lattice=lattice)
 
     def __mul__(self, m, method="tile") -> Geometry:
         """Implement easy tile/repeat function
@@ -2417,122 +2210,6 @@ class Geometry(
             return ang
         return np.degrees(ang)
 
-    @deprecate_argument(
-        "only",
-        "what",
-        "argument only has been deprecated in favor of what, please update your code.",
-        "0.14.0",
-    )
-    def rotate(
-        self,
-        angle: float,
-        v,
-        origin=None,
-        atoms: Optional[AtomsArgument] = None,
-        rad: bool = False,
-        what: Optional[str] = None,
-    ) -> Geometry:
-        r"""Rotate geometry around vector and return a new geometry
-
-        Per default will the entire geometry be rotated, such that everything
-        is aligned as before rotation.
-
-        However, by supplying ``what = 'abc|xyz'`` one can designate which
-        part of the geometry that will be rotated.
-
-        Parameters
-        ----------
-        angle : float
-             the angle in degrees to rotate the geometry. Set the ``rad``
-             argument to use radians.
-        v     : int or str or array_like
-             the normal vector to the rotated plane, i.e.
-             v = [1,0,0] will rotate around the :math:`yz` plane.
-             If a str it refers to the Cartesian direction (xyz), or the
-             lattice vectors (abc). Providing several is the combined direction.
-        origin : int or array_like, optional
-             the origin of rotation. Anything but [0, 0, 0] is equivalent
-             to a `self.move(-origin).rotate(...).move(origin)`.
-             If this is an `int` it corresponds to the atomic index.
-        atoms : int or array_like, optional
-             only rotate the given atomic indices, if not specified, all
-             atoms will be rotated.
-        rad : bool, optional
-             if ``True`` the angle is provided in radians (rather than degrees)
-        what : {'xyz', 'abc', 'abc+xyz', <or combinations of "xyzabc">}
-            which coordinate subject should be rotated,
-            if any of ``abc`` is in this string the corresponding cell vector will be rotated
-            if any of ``xyz`` is in this string the corresponding coordinates will be rotated
-            If `atoms` is None, this defaults to "abc+xyz", otherwise it defaults
-            to "xyz". See Examples.
-
-        Examples
-        --------
-        rotate coordinates around the :math:`x`-axis
-        >>> geom_x45 = geom.rotate(45, [1, 0, 0])
-
-        rotate around the ``(1, 1, 0)`` direction but project the rotation onto the :math:`x`
-        axis
-        >>> geom_xy_x = geom.rotate(45, "xy", what='x')
-
-        See Also
-        --------
-        Quaternion : class to rotate
-        Lattice.rotate : rotation passed to the contained supercell
-        """
-        if origin is None:
-            origin = [0.0, 0.0, 0.0]
-        elif isinstance(origin, Integral):
-            origin = self.axyz(origin)
-        origin = _a.asarray(origin)
-
-        if atoms is None:
-            if what is None:
-                what = "abc+xyz"
-            # important to not add a new dimension to xyz
-            atoms = slice(None)
-        else:
-            if what is None:
-                what = "xyz"
-            # Only rotate the unique values
-            atoms = self.sc2uc(atoms, unique=True)
-
-        if isinstance(v, Integral):
-            v = direction(v, abc=self.cell, xyz=np.diag([1, 1, 1]))
-        elif isinstance(v, str):
-            v = reduce(
-                lambda a, b: a + direction(b, abc=self.cell, xyz=np.diag([1, 1, 1])),
-                v,
-                0,
-            )
-
-        # Ensure the normal vector is normalized... (flatten == copy)
-        vn = _a.asarrayd(v).flatten()
-        vn /= fnorm(vn)
-
-        # Rotate by direct call
-        lattice = self.lattice.rotate(angle, vn, rad=rad, what=what)
-
-        # Copy
-        xyz = np.copy(self.xyz)
-
-        idx = []
-        for i, d in enumerate("xyz"):
-            if d in what:
-                idx.append(i)
-
-        if idx:
-            # Prepare quaternion...
-            q = Quaternion(angle, vn, rad=rad)
-            q /= q.norm()
-            # subtract and add origin, before and after rotation
-            rotated = q.rotate(xyz[atoms] - origin) + origin
-            # get which coordinates to rotate
-            for i in idx:
-                xyz[atoms, i] = rotated[:, i]
-
-        return self.__class__(xyz, atoms=self.atoms.copy(), lattice=lattice)
-
     def rotate_miller(self, m, v) -> Geometry:
         """Align Miller direction along ``v``
 
@@ -2635,352 +2312,6 @@ class Geometry(
         return self.__class__(
             xyz, atoms=self.atoms.swap(atoms_a, atoms_b), lattice=self.lattice.copy()
         )
-
-    def swapaxes(
-        self, axes_a: Union[int, str], axes_b: Union[int, str], what: str = "abc"
-    ) -> Geometry:
-        """Swap the axes components by either lattice vectors (only cell), or Cartesian coordinates
-
-        See `Lattice.swapaxes` for details.
-
-        Parameters
-        ----------
-        axes_a :
-           the old axis indices (or labels if `str`)
-           A string will translate each character as a specific
-           axis index.
-           Lattice vectors are denoted by ``abc`` while the
-           Cartesian coordinates are denote by ``xyz``.
-           If `str`, then `what` is not used.
-        axes_b :
-           the new axis indices, same as `axes_a`
-           old axis indices (or labels)
-        what : {'abc', 'xyz', 'abc+xyz'}
-           what to swap, lattice vectors (abc) or Cartesian components (xyz),
-           or both.
-           Neglected for integer axes arguments.
-
-        See Also
-        --------
-        Lattice.swapaxes
-
-        Examples
-        --------
-
-        Only swap lattice vectors
-
-        >>> g_ba = g.swapaxes(0, 1)
-        >>> assert np.allclose(g.xyz, g_ba.xyz)
-
-        Only swap Cartesian coordinates
-
-        >>> g_ba = g.swapaxes(0, 1, "xyz")
-        >>> assert np.allclose(g.xyz[:, [1, 0, 2]], g_ba.xyz)
-
-        Consecutive swappings (what will be neglected if provided):
-
-        1. abc, xyz -> bac, xyz
-        2. bac, xyz -> bca, xyz
-        3. bac, xyz -> bca, zyx
-
-        >>> g_s = g.swapaxes("abx", "bcz")
-        >>> assert np.allclose(g.xyz[:, [2, 1, 0]], g_s.xyz)
-        >>> assert np.allclose(g.cell[[1, 2, 0]][:, [2, 1, 0]], g_s.cell)
-        """
-        # swap supercell
-        # We do not need to check argument types etc,
-        # Lattice.swapaxes will do this for us
-        lattice = self.lattice.swapaxes(axes_a, axes_b, what)
-
-        if isinstance(axes_a, int) and isinstance(axes_b, int):
-            if "xyz" in what:
-                axes_a = "xyz"[axes_a]
-                axes_b = "xyz"[axes_b]
-            else:
-                axes_a = ""
-                axes_b = ""
-
-        # only thing we are going to swap is the coordinates
-        idx = [0, 1, 2]
-        for a, b in zip(axes_a, axes_b):
-            aidx = "xyzabc".index(a)
-            bidx = "xyzabc".index(b)
-            if aidx < 3:
-                idx[aidx], idx[bidx] = idx[bidx], idx[aidx]
-
-        return self.__class__(
-            self.xyz[:, idx].copy(), atoms=self.atoms.copy(), lattice=lattice
-        )
-
-    def center(
-        self, atoms: Optional[AtomsArgument] = None, what: str = "xyz"
-    ) -> ndarray:
-        """Returns the center of the geometry
-
-        By specifying `what` one can control whether it should be:
-
-        * ``cop|xyz|position``: Center of coordinates (default)
-        * ``mm:xyz`` or ``mm(xyz)``: Center of minimum/maximum of coordinates
-        * ``com|mass``: Center of mass
-        * ``com:pbc|mass:pbc``: Center of mass using periodicity, if the point 0, 0, 0 is returned it
-            may likely be because of a completely periodic system with no true center of mass
-        * ``cou|cell``: Center of cell
-
-        Parameters
-        ----------
-        atoms :
-            list of atomic indices to find center of
-        what : {'xyz', 'mm:xyz', 'mass', 'mass:pbc', 'cell'}
-            determine which center to calculate
-        """
-        what = what.lower()
-        if what in ("cou", "cell", "lattice"):
-            return self.lattice.center()
-
-        if atoms is None:
-            g = self
-        else:
-            g = self.sub(atoms)
-
-        if what in ("com:pbc", "mass:pbc"):
-            mass = g.mass
-            sum_mass = mass.sum()
-            # the periodic center of mass is determined by transfering all
-            # coordinates onto a circle -> fxyz * 2pi
-            # Then we mass average the circle angles for each of the fractional
-            # coordinates, and transform back into the cartesian coordinate system
-            theta = g.fxyz * (2 * np.pi)
-            # construct angles
-            avg_cos = (mass @ np.cos(theta)) / sum_mass
-            avg_sin = (mass @ np.sin(theta)) / sum_mass
-            avg_theta = np.arctan2(-avg_sin, -avg_cos) / (2 * np.pi) + 0.5
-            return avg_theta @ g.lattice.cell
-
-        if what in ("com", "mass"):
-            mass = g.mass
-            return mass @ g.xyz / mass.sum()
-
-        if what in ("mm:xyz", "mm(xyz)"):
-            return (g.xyz.min(0) + g.xyz.max(0)) / 2
-
-        if what in ("cop", "xyz", "position"):
-            return np.mean(g.xyz, axis=0)
-
-        raise ValueError(
-            f"{self.__class__.__name__}.center could not understand option 'what' got {what}"
-        )
-
-    def append(
-        self, other: LatticeOrGeometryLike, axis: int, offset="none"
-    ) -> Geometry:
-        """Appends two structures along `axis`
-
-        This will automatically add the ``self.cell[axis,:]`` to all atomic
-        coordiates in the `other` structure before appending.
-
-        The basic algorithm is this:
-
-        >>> oxa = other.xyz + self.cell[axis,:][None,:]
-        >>> self.xyz = np.append(self.xyz,oxa)
-        >>> self.cell[axis,:] += other.cell[axis,:]
-
-        NOTE: The cell appended is only in the axis that
-        is appended, which means that the other cell directions
-        need not conform.
-
-        Parameters
-        ----------
-        other :
-            Other geometry class which needs to be appended
-            If a `Lattice` only the super cell will be extended
-        axis :
-            Cell direction to which the `other` geometry should be
-            appended.
-        offset : {'none', 'min', (3,)}
-            By default appending two structures will simply use the coordinates,
-            as is.
-            With 'min', the routine will shift both the structures along the cell
-            axis of `self` such that they coincide at the first atom, lastly one
-            may use a specified offset to manually select how `other` is displaced.
-            NOTE: That `self.cell[axis, :]` will be added to `offset` if `other` is
-            a geometry.
-
-        See Also
-        --------
-        add : add geometries
-        prepend : prending geometries
-        attach : attach a geometry
-        insert : insert a geometry
-        """
-        if isinstance(other, Lattice):
-            # Only extend the supercell.
-            xyz = np.copy(self.xyz)
-            atoms = self.atoms.copy()
-            lattice = self.lattice.append(other, axis)
-            names = self._names.copy()
-            if isinstance(offset, str):
-                if offset == "none":
-                    offset = [0, 0, 0]
-                else:
-                    raise ValueError(
-                        f"{self.__class__.__name__}.append requires offset to be (3,) for supercell input"
-                    )
-            xyz += _a.asarray(offset)
-
-        else:
-            # sanitize output
-            other = self.new(other)
-            if isinstance(offset, str):
-                offset = offset.lower()
-                if offset == "none":
-                    offset = self.cell[axis, :]
-                elif offset == "min":
-                    # We want to align at the minimum position along the `axis`
-                    min_f = self.fxyz[:, axis].min()
-                    min_other_f = dot(other.xyz, self.icell.T)[:, axis].min()
-                    offset = self.cell[axis, :] * (1 + min_f - min_other_f)
-                else:
-                    raise ValueError(
-                        f"{self.__class__.__name__}.append requires align keyword to be one of [none, min, (3,)]"
-                    )
-            else:
-                offset = self.cell[axis, :] + _a.asarray(offset)
-
-            xyz = np.append(self.xyz, offset + other.xyz, axis=0)
-            atoms = self.atoms.append(other.atoms)
-            lattice = self.lattice.append(other.lattice, axis)
-            names = self._names.merge(other._names, offset=len(self))
-
-        return self.__class__(xyz, atoms=atoms, lattice=lattice, names=names)
-
-    def prepend(
-        self, other: LatticeOrGeometryLike, axis: int, offset="none"
-    ) -> Geometry:
-        """Prepend two structures along `axis`
-
-        This will automatically add the ``self.cell[axis,:]`` to all atomic
-        coordiates in the `other` structure before appending.
-
-        The basic algorithm is this:
-
-        >>> oxa = other.xyz
-        >>> self.xyz = np.append(oxa, self.xyz + other.cell[axis,:][None,:])
-        >>> self.cell[axis,:] += other.cell[axis,:]
-
-        NOTE: The cell prepended is only in the axis that
-        is prependend, which means that the other cell directions
-        need not conform.
-
-        Parameters
-        ----------
-        other :
-            Other geometry class which needs to be prepended
-            If a `Lattice` only the super cell will be extended
-        axis :
-            Cell direction to which the `other` geometry should be
-            prepended
-        offset : {'none', 'min', (3,)}
-            By default appending two structures will simply use the coordinates,
-            as is.
-            With 'min', the routine will shift both the structures along the cell
-            axis of `other` such that they coincide at the first atom, lastly one
-            may use a specified offset to manually select how `self` is displaced.
-            NOTE: That `other.cell[axis, :]` will be added to `offset` if `other` is
-            a geometry.
-
-        See Also
-        --------
-        add : add geometries
-        append : appending geometries
-        attach : attach a geometry
-        insert : insert a geometry
-        """
-        if isinstance(other, Lattice):
-            # Only extend the supercell.
-            xyz = np.copy(self.xyz)
-            atoms = self.atoms.copy()
-            lattice = self.lattice.prepend(other, axis)
-            names = self._names.copy()
-            if isinstance(offset, str):
-                if offset == "none":
-                    offset = [0, 0, 0]
-                else:
-                    raise ValueError(
-                        f"{self.__class__.__name__}.prepend requires offset to be (3,) for supercell input"
-                    )
-            xyz += _a.arrayd(offset)
-
-        else:
-            # sanitize output
-            other = self.new(other)
-            if isinstance(offset, str):
-                offset = offset.lower()
-                if offset == "none":
-                    offset = other.cell[axis, :]
-                elif offset == "min":
-                    # We want to align at the minimum position along the `axis`
-                    min_f = other.fxyz[:, axis].min()
-                    min_other_f = dot(self.xyz, other.icell.T)[:, axis].min()
-                    offset = other.cell[axis, :] * (1 + min_f - min_other_f)
-                else:
-                    raise ValueError(
-                        f"{self.__class__.__name__}.prepend requires align keyword to be one of [none, min, (3,)]"
-                    )
-            else:
-                offset = other.cell[axis, :] + _a.asarray(offset)
-
-            xyz = np.append(other.xyz, offset + self.xyz, axis=0)
-            atoms = self.atoms.prepend(other.atoms)
-            lattice = self.lattice.prepend(other.lattice, axis)
-            names = other._names.merge(self._names, offset=len(other))
-
-        return self.__class__(xyz, atoms=atoms, lattice=lattice, names=names)
-
-    def add(
-        self, other: LatticeOrGeometryLike, offset: Sequence[float] = (0, 0, 0)
-    ) -> Geometry:
-        """Merge two geometries (or a Geometry and Lattice) by adding the two atoms together
-
-        If `other` is a Geometry only the atoms gets added, to also add the supercell vectors
-        simply do ``geom.add(other).add(other.lattice)``.
-
-        Parameters
-        ----------
-        other :
-            Other geometry class which is added
-        offset :
-            offset in geometry of `other` when adding the atoms.
-            Otherwise it is the offset of the `self` atoms.
-
-        See Also
-        --------
-        append : appending geometries
-        prepend : prending geometries
-        attach : attach a geometry
-        insert : insert a geometry
-
-        Examples
-        --------
-        >>> first = Geometry(...)
-        >>> second = Geometry(...)
-        >>> lattice = Lattice(...)
-        >>> added = first.add(second, offset=(0, 0, 2))
-        >>> assert np.allclose(added.xyz[:len(first)], first.xyz)
-        >>> assert np.allclose(added.xyz[len(first):] - [0, 0, 2], second.xyz)
-
-        """
-        if isinstance(other, Lattice):
-            xyz = self.xyz.copy() + _a.arrayd(offset)
-            lattice = self.lattice + other
-            atoms = self.atoms.copy()
-            names = self._names.copy()
-        else:
-            other = self.new(other)
-            xyz = np.append(self.xyz, other.xyz + _a.arrayd(offset), axis=0)
-            lattice = self.lattice.copy()
-            atoms = self.atoms.add(other.atoms)
-            names = self._names.merge(other._names, offset=len(self))
-        return self.__class__(xyz, atoms=atoms, lattice=lattice, names=names)
 
     def add_vacuum(
         self, vacuum: float, axis: int, offset: Sequence[float] = (0, 0, 0)
