+++ src/sisl/sparse_geometry.py
@@ -22,19 +22,13 @@ from numpy import (
     lexsort,
     repeat,
     searchsorted,
-    take,
     tile,
     unique,
 )
 from numpy.lib.mixins import NDArrayOperatorsMixin
 from scipy.sparse import csr_matrix
 
-if TYPE_CHECKING:
-    from sisl.typing import AtomsArgument
-
-from sisl._core.atom import Atom
-from sisl._core.geometry import Geometry
-from sisl._core.orbital import Orbital
+from sisl._core import Atom, Geometry, Orbital
 
 from . import _array as _a
 from ._array import array_arange
@@ -42,6 +36,9 @@ from ._internal import set_module
 from .messages import SislError, SislWarning, progressbar, warn
 from .sparse import SparseCSR, _ncol_to_indptr, issparse
 from .utils.ranges import list2str
+
+if TYPE_CHECKING:
+    from sisl.typing import AtomsArgument
 
 __all__ = ["SparseAtom", "SparseOrbital"]
